Changed namespace of ElectionFactory and ElectionResponce
